+++ pastry/src/rice/p2p/replication/manager/messaging/TimeoutMessage.java
@@ -39,7 +39,7 @@ public class TimeoutMessage implements Message {
    *
    * @return This message's priority
    */
-  public int getPriority() {
+  public byte getPriority() {
     return MEDIUM_PRIORITY;
   }
